Fixed some bugs, added some features
Added a console printing when an error occurs to go along with the text
log.
Added dates in error log so that you know when it occured.
Separated checkForInternet() to its own method.
The temporary date file used to check the date will now be stored in the
temporary directory rather than the "user.dir".
Changed settings name to gamepatchersettings.txt and made settings more
clear to understand.