Allowed users allowing updates or not
Now a setDenial(boolean) method in Downloader will allow for you to set
the program to confirm that the user wants to update.  You should either
call setDenial(true) in your main method if you never want the program
to auto-update (denial = false) or make an argument checker like I did
in Example.  The default is false (auto-update) but you can easily
switch it to true with setDenial(true).

Something that I'm not sure about though is what to default denial to.
Is auto-update by default or no auto-update by default good?