Got rid of pinging
Pinging is not needed, as if the connection to the file site fails, then
there is obviously no internet.  Also made the patcher easier to use.
You can still just use downloader.autoUpdate() (used to be
checkForUpdate()), but now you can also manually check for an update
then manually update if you prefer.  That's what I do in my example,
which is now updated.